requirements.txt
crispy_bootstrap4==2022.1
Django==4.2.5
django-cas-ng==5.0.0
django-crispy-forms==2.0
django-redis==5.3.0
LightPipes==2.1.4
numpy==1.26.0
pandas==2.1.1
plotly==5.17.0
python-environ==0.4.54
python-ldap===3.4.3
redis==5.0.0
secret-key-generator==0.0.8
selph-light-lib==0.0.5
